+++ components.json
@@ -11,9 +11,9 @@
         "prefix": ""
     },
     "aliases": {
-        "components": "@/components",
-        "utils": "@/lib/utils",
-        "ui": "@/components/ui",
+        "components": "@/shared/ui",
+        "utils": "@/shared/lib/utils",
+        "ui": "@/shared/ui",
         "lib": "@/lib",
         "hooks": "@/hooks"
     },
