+++ src/shared/ui/drawer.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 
 import { Drawer as DrawerPrimitive } from "vaul";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 const Drawer = ({
     shouldScaleBackground = true,
