+++ src/shared/ui/slider.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 import * as SliderPrimitive from "@radix-ui/react-slider";
 
 const Slider = React.forwardRef<
