+++ src/shared/ui/toggle.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 import * as TogglePrimitive from "@radix-ui/react-toggle";
 
 const toggleVariants = cva(
